Added depth_profile plot function and cleaned up other functions;

+++ analyse_trajectory.py
@@ -17,6 +17,7 @@ class Analyse:
         self.transit_file = f.save + f.sim + '/'
         self.dom_file = f.save + f.sim + '/dominant_paths.npy'  # save dominant paths
         self.depth_file = f.save + f.sim + '/depth_profile.npy'  # save dominant paths
+        self.retention_file = f.save + f.sim + '/retention.npy'  # save dominant paths
         return
 
     def dom_paths(self, k):
@@ -36,23 +37,21 @@ class Analyse:
             return
 
     def depth_profile(self, k):
-        max_depth = 200
-        z_bins = np.zeros([max_depth, k.t_max])
-        #  Find and loop over batches of individuals
-        for t_id in range(0, np.shape(k.x)[0]):
-            z = k.get_slice(t_id)
-            z_ik = np.floor(z[:])
-            for i in range(1, max_depth):
-                z_bins[i, t_id] = np.sum(i == z_ik)
+        if not os.path.exists(self.depth_file):
+            max_depth = 200
+            z_bins = np.zeros([max_depth, k.t_max])
+            #  Find and loop over batches of individuals
+            for t_id in range(0, np.shape(k.x)[0]):
+                z = k.get_slice_z(t_id)
+                z_ik = np.floor(z[:])
+                for i in range(1, max_depth):
+                    z_bins[i, t_id] = np.sum(i == z_ik)
 
-        np.save(self.depth_file, z_bins)
+            np.save(self.depth_file, z_bins)
+        else:
+            print('Directory: ' + self.depth_file + ' already exists, skipping')
         return
 
-        # import matplotlib.pyplot as plt
-        # f1 = plt.contourf(z_bins)
-        # ax = f1.axes
-        # ax.invert_yaxis()
-        # breakpoint()
 
     def transit_times(self, key, k):
         self.ssmu_target(key)
@@ -84,74 +83,44 @@ class Analyse:
                 print('Directory: ' + self.transit_file + ' already exists, skipping')
         return
 
+    def retention_times(self, k):
+        if not os.path.exists(self.retention_file):
+            df = np.zeros(np.shape(k.depth))
+            df[np.isnan(k.depth)] = np.nan
+            act_part = k.act_part[:]
+            time = k.time
+            d_lim = 5
+            temp_vec = np.zeros(np.shape(k.x)[0])
+            x = np.array(k.x[:])
+            y = np.array(k.y[:])
+            store_mat = np.array([x[0:np.sum(act_part==0),0], y[0:np.sum(act_part==0),0], np.zeros(np.shape(y[0:np.sum(act_part==0), 0]))]).T
+            for i in range(0, np.shape(k.x)[0]):
+                act_id = act_part[i]
+                xt = x[i, act_id:np.shape(x)[1]].astype(float)
+                yt = y[i, act_id:np.shape(y)[1]].astype(float)
+                y_dist = (yt - yt[0])**2
+                x_dist = (xt - xt[0])**2
+                #y_dist[y_dist <= 0] = 0.0001
+                #x_dist[x_dist <= 0] = 0.0001
+                dist_t = np.sqrt(x_dist + y_dist)
+                cond = np.where(dist_t > d_lim)
+                if np.shape(cond)[0]*np.shape(cond)[1] > 0:
+                    lim_v = cond[0][0] + act_id
+                    date_0 = time[act_id]  # time particle became active
+                    date_1 = time[lim_v]  # time particle reached the target destination;
+                    timeframe = date_1 - date_0
+                    transit_hours = (timeframe.days * 24) + np.floor(timeframe.seconds * 1 / (60 * 60))
+                    temp_vec[i] = transit_hours
+                else:
+                    temp_vec[i] = np.nan
 
-
-
-
-
-                #else
-
-                #todo: save sites instead of all particles;
-
-
-
-
-    #     subs = ssmu_target(sub_key)  # Target regions stored in dictionary (there may be multiple target areas);
-    #     for target_area in subs:
-    #         filepath = list_dir[sub_key + '_folder'] + target_area + '_' + 'transit.npy'
-    #         filepath2 = list_dir[sub_key + '_folder'] + target_area + '_' + 'wormx.npy'
-    #         filepath3 = list_dir[sub_key + '_folder'] + target_area + '_' + 'wormy.npy'
-    #         filepath4 = list_dir[sub_key + '_folder'] + target_area + '_' + 'act_part.npy'
-    #         if not os.path.exists(filepath):
-    #             # Load depth file and region file:
-    #             depth = np.load(list_dir['depth_file'])
-    #             nc_file = nc.Dataset(list_dir['reg_file'], mode='r', format='NETCDF4_CLASSIC')
-    #             df = np.zeros(np.shape(depth))
-    #             df[np.isnan(depth)] = np.nan
-    #             x = nc_file['xp'][idv, :]
-    #             y = nc_file['yp'][idv, :]
-    #             in_reg = nc_file['in_region'][idv, :]
-    #             act_part = nc_file['act_part'][idv]
-    #             time = np.load(list_dir['time_file'], allow_pickle=True)
-    #             transit_mat = np.zeros([np.shape(x)[0], 5])
-    #             n_saves = 2500
-    #             worm_matx = np.empty([n_saves, np.shape(x)[1]])
-    #             worm_maty = np.empty([n_saves, np.shape(x)[1]])
-    #             sub_ids = subs[target_area]
-    #             print(subs[target_area])
-    #             c = -1
-    #             for i in range(0, np.shape(x)[0]):
-    #                 act_id = act_part[i]  # when particle becomes active
-    #                 visit_reg = in_reg[i, :]  # All the regions particle visits
-    #                 if not np.isin(in_reg[i, act_id], sub_ids):  # Make sure it doesn't start in region
-    #                     ids = np.where(np.isin(visit_reg, sub_ids))  # where there are overlaps
-    #                     if not np.shape(ids)[0]*np.shape(ids)[1] == 0:
-    #                         transit_mat[i, 0] = x[i, act_id]
-    #                         transit_mat[i, 1] = y[i, act_id]
-    #                         date_0 = time[act_id]  # time particle became active
-    #                         date_1 = time[ids[0][0]]  # time particle reached the target destination;
-    #                         timeframe = date_1 - date_0
-    #                         transit_hours = (timeframe.days*24) + np.floor(timeframe.seconds*1/(60*60))
-    #                         transit_mat[i, 2] = transit_hours
-    #                         transit_mat[i, 3] = x[i, ids[0][0]]
-    #                         transit_mat[i, 4] = y[i, ids[0][0]]
-    #                         #Worm matrix
-    #                         if not c == n_saves:
-    #                             worm_matx[c, act_id:np.shape(x)[1]] = x[i, act_id:np.shape(x)[1]]
-    #                             worm_maty[c, act_id:np.shape(x)[1]] = y[i, act_id:np.shape(x)[1]]
-    #                             c = c + 1
-    #
-    #                     # Add column with index of arrival perhaps
-    #             print('Saving: ' + filepath)
-    #             np.save(filepath, transit_mat)
-    #             np.save(filepath2, worm_matx)
-    #             np.save(filepath3, worm_maty)
-    #             # when particles are active:
-    #             nc_file.close()  # close nc file
-    #         else:
-    #             print('Directory: ' + filepath + ' already exists, skipping')
-    #     return
-    #
+            for i in range(0, np.shape(store_mat)[0]):
+                ids = np.arange(i, np.shape(x)[0], np.shape(store_mat)[0])
+                store_mat[i, 2] = np.nanmean(temp_vec[ids])
+            np.save(self.retention_file, store_mat)
+        else:
+            print('Directory: ' + self.retention_file + ' already exists, skipping')
+        return
 
     def ssmu_target(self, key):
         # Hard-coded indices for particles starting in ssmu's. This can be adapted for both time and region;
@@ -230,42 +199,7 @@ class Analyse:
 #     nc_file.close()
 #     return
 #
-# def retention_times(idv, list_dir, sub_key):
-#     depth = np.load(list_dir['depth_file'])
-#     nc_file = nc.Dataset(list_dir['reg_file'], mode='r', format='NETCDF4_CLASSIC')
-#     df = np.zeros(np.shape(depth))
-#     df[np.isnan(depth)] = np.nan
-#     x = nc_file['xp'][idv, :].astype(int)
-#     y = nc_file['yp'][idv, :].astype(int)
-#     act_part = nc_file['act_part'][idv]
-#     store_mat = np.array([x[0:np.sum(act_part==0),0], y[0:np.sum(act_part==0), 0], np.zeros(np.shape(y[0:np.sum(act_part==0), 0]))]).T
-#     time = np.load(list_dir['time_file'], allow_pickle=True)
-#     d_lim = 5
-#     temp_vec = np.zeros(np.shape(x)[0])
-#     for i in range(0, np.shape(x)[0]):
-#         act_id = act_part[i]
-#         xt = x[i, act_id:np.shape(x)[1]]
-#         yt = y[i, act_id:np.shape(x)[1]]
-#         dist_t = np.sqrt(((xt-xt[0])**2) + ((yt-yt[0])**2))
-#         cond = np.where(dist_t > d_lim)
-#         if np.shape(cond)[0]*np.shape(cond)[1] > 0:
-#             lim_v = cond[0][0] + act_id
-#             date_0 = time[act_id]  # time particle became active
-#             date_1 = time[lim_v]  # time particle reached the target destination;
-#             timeframe = date_1 - date_0
-#             transit_hours = (timeframe.days * 24) + np.floor(timeframe.seconds * 1 / (60 * 60))
-#             temp_vec[i] = transit_hours
-#         else:
-#             temp_vec[i] = np.nan
-#
-#     for i in range(0, np.shape(store_mat)[0]):
-#         ids = np.arange(i,np.shape(x)[0], np.shape(store_mat)[0])
-#         store_mat[i, 2] = np.nanmean(temp_vec[ids])
-#
-#     save_path = list_dir[sub_key + '_folder'] + 'retention.npy'
-#     np.save(save_path, store_mat)
-#     nc_file.close()
-#     return
+
 #
 #
